Refactor actions into smaller files
To make space for the substantially more complex "drain"
action, so things don't get so cluttered.
